requirements.txt
streamlit==1.28.2
python-dotenv==1.0.0
faiss-cpu==1.9.0
numpy==1.26.4
scikit-learn==1.3.2
google-generativeai==0.3.2
